Using fms configurator to create a fms with a resource manager
