+++ mrs/robot.py
@@ -38,7 +38,7 @@ if __name__ == '__main__':
     from fleet_management.config.loader import Configurator
 
     config_file_path = '../config/config.yaml'
-    config = Configurator(config_file_path, initialize=False)
+    config = Configurator(config_file_path)
     config.configure_logger()
 
     parser = argparse.ArgumentParser()
